requirements.txt
requests
python-telegram-bot==13.8.1
tqdm==4.62.3
